fix: rename user-prisma service
